[hf-card] Refactorted code for simplicity and resuse. created first class decorator that injects other card decorators.
