+++ metaflow_extensions/deepspeed/plugins/deepspeed_libs/hugging_face/constants.py
@@ -1,0 +1,1 @@
+DEFAULT_FILE_NAME = "huggingface_model_card.json"
