+++ metaflow_extensions/deepspeed/plugins/mfextinit_deepspeed.py
@@ -1,4 +1,5 @@
 STEP_DECORATORS_DESC = [
-    ("deepspeed", ".deepspeed_libs.deepspeed_decorator.DeepspeedDecorator")
+    ("deepspeed", ".deepspeed_libs.deepspeed_decorator.DeepspeedDecorator"),
+    ("huggingface_card", ".deepspeed_libs.hugging_face.card_decorator.HuggingFaceCardDecorator"),
 ]
-__mf_promote_submodules__ = ["az_store", "hf_callbacks", "deepspeed_libs", "hf_info_card"]
+__mf_promote_submodules__ = ["az_store", "hf_callbacks", "deepspeed_libs",]
