+++ setup.py
@@ -1,6 +1,6 @@
 from setuptools import setup, find_namespace_packages
 
-version = "0.0.7"
+version = "0.0.9"
 
 setup(
     name="metaflow-deepspeed",
